apps/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
pydantic-settings
pgvector
pypdf
python-docx
openai
alembic
